Refactored the fetch function to the menulist
